- NOTE: The paths that I have in here (specifically in age_racer.py) may not be fully correct and you should adjust them for your own machine. This is unfortunately the only way it works on mine, and I don't know what is the correct way to make it work for everyone.
Essentially just made a more complete agent. It may not be the final result, since perhaps we may be interested in altering more parameters, but it's very close. The results are quite positive, and after only running it a couple of times with 50 generations it reduced the time from 432 to 421.