Links to DESMOS graphs for agent functionality
Here are links to the graphs I created in DESMOS to help me model the parameters and how they combine with one another to produce the results that were given. Play around with the parameter values as you wish, and even change the ranges. It should help with understanding their functionality.